pyproject.toml
[tool.poetry]
name = "vlm-eval"
version = "0.1.0"
description = "A modular, scenario-driven evaluation framework for Vision-Language Models"
authors = ["Bashir Alam <bashir.alam@abo.fi>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/bashiralam185/vlm-eval-framework"
repository = "https://github.com/bashiralam185/vlm-eval-framework"
keywords = ["vlm", "vision-language", "evaluation", "benchmarking", "multimodal", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "vlm_eval"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2"
torchvision = "^0.17"
transformers = "^4.40"
Pillow = "^10.3"
numpy = "^1.26"
pandas = "^2.1"
pydantic = "^2.7"
pydantic-settings = "^2.2"
pyyaml = "^6.0"
loguru = "^0.7"
rich = "^13.7"
typer = "^0.12"
fastapi = "^0.111"
uvicorn = {extras = ["standard"], version = "^0.29"}
httpx = "^0.27"
plotly = "^5.22"
jinja2 = "^3.1"
nltk = "^3.8"
scikit-learn = "^1.4"
tqdm = "^4.66"
requests = "^2.31"
pycocotools = "^2.0"
accelerate = "^0.30"
einops = "^0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-cov = "^5.0"
pytest-asyncio = "^0.23"
ruff = "^0.4"
mypy = "^1.10"
pre-commit = "^3.7"

[tool.poetry.scripts]
vlm-eval = "vlm_eval.cli:app"
vlm-api = "api.main:run_server"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "B", "UP"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=vlm_eval --cov-report=term-missing --cov-report=xml"
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]
